[fix](infra): fix number type issue with big and small numbers
t# Please enter the commit message for your changes. Lines starting

+++ src/runtime/src/value.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone)]
 pub enum Value {
-    Number(i128),
+    Number(f64),
     String(String),
     Bool(bool),
     Null,
